Cleanup and move test for obsolete code.
