+++ tests/test_snr_normalization.py
@@ -1,38 +1,23 @@
-import os
-
 import numpy as np
 import pytest
 
-import eniric
-import eniric.IOmodule as Io
-import eniric.Qcalculator as Q
 import eniric.snr_normalization as snrnorm
 import eniric.utilities as utils
-
-resampled_template = "Spectrum_{0}-PHOENIX-ACES_{1}band_vsini{2}_R{3}_res3.0.txt"
-wave_photon_template = (
-    "lte0{0}-4.50-0.0.PHOENIX-ACES-AGSS-COND-2011-HiRes_wave_photon.dat"
-)
+from eniric_scripts.aces_precision import convolve_and_resample
 
 
-@pytest.mark.parametrize("temp", [2800, 2600])
 @pytest.mark.parametrize("desired_snr", [100.0, 150.0])
 @pytest.mark.parametrize("band", ["J", "Y", "VIS"])
-def test_snr_normalization(desired_snr, band, temp):
+def test_snr_normalization(desired_snr, band, testing_spectrum):
     """Test SNR after normalizing function is the desired value.
 
     Testing on middle of J band.
     """
-    test_data = os.path.join(
-        eniric.paths["phoenix_dat"], "Z-0.0", wave_photon_template.format(temp)
-    )
-
+    wav, flux = testing_spectrum
     band_mid = utils.band_middle(band)
-    wav, flux = utils.read_spectrum(test_data)
 
-    index_reference = np.searchsorted(wav, [band_mid])[
-        0
-    ]  # Searching for the closest index to 1.25
+    # Searching for the closest index to 1.25
+    index_reference = np.searchsorted(wav, [band_mid])[0]
     snr_estimate = np.sqrt(np.sum(flux[index_reference - 1 : index_reference + 2]))
 
     assert round(snr_estimate, 1) != desired_snr  # Assert SNR is not correct
@@ -47,34 +32,24 @@ def test_snr_normalization(desired_snr, band, temp):
     assert round(new_snr_estimate, 0) == desired_snr
 
 
-@pytest.mark.parametrize("temp", [3900, 3500])
 @pytest.mark.parametrize("desired_snr", [50.0, 200.0])
 @pytest.mark.parametrize("band", ["H", "GAP", "K"])
-def test_snr_normalization_constant(desired_snr, band, temp):
+def test_snr_normalization_constant(desired_snr, band, testing_spectrum):
     """Test snr_constant_band and snr_constant_wav produce same result."""
-    test_data = os.path.join(
-        eniric.paths["phoenix_dat"], "Z-0.0", wave_photon_template.format(temp)
-    )
-
+    wav, flux = testing_spectrum
     band_mid = utils.band_middle(band)
-    wav, flux = utils.read_spectrum(test_data)
 
     assert snrnorm.snr_constant_band(
         wav, flux, band=band, snr=desired_snr
     ) == snrnorm.snr_constant_wav(wav, flux, band_mid, snr=desired_snr)
 
 
-def test_band_snr_norm():
+def test_band_snr_norm(testing_spectrum):
     """Compared to wav snr norm."""
-    # snr_constant_band
-    star, band, vel, res = "M0", "J", 1.0, "100k"
-    test_data = os.path.join(
-        eniric.paths["test_data"],
-        "resampled",
-        resampled_template.format(star, band, vel, res),
+    wav, flux = testing_spectrum
+    wav, flux = convolve_and_resample(
+        wav, flux, vsini=1, R=100000, band="J", sampling=3
     )
-    wav, flux = Io.pdread_2col(test_data)
-
     assert snrnorm.snr_constant_band(
         wav, flux, band="J", snr=100
     ) == snrnorm.snr_constant_wav(wav, flux, wav_ref=1.25, snr=100)
@@ -119,132 +94,6 @@ def test_sampling_index_array():
         snrnorm.sampling_index(
             46, 10, array_length=50
         )  # an index will be > (array_length - 1)
-
-
-@pytest.mark.parametrize("bad_string", ["id-string", "M0-K-1.0-100", "M0-P-1.0-100k"])
-def test_errors_in_snr_get_reference_spectrum(bad_string):
-    """Testing Errors in getting the reference spectrum."""
-    with pytest.raises(ValueError):
-        snrnorm.get_reference_spectrum(bad_string)
-
-
-@pytest.mark.parametrize("bad_string", ["Alpha=", "smpl="])
-def test_notimplemented_errors_in_snr_get_reference_spectrum(bad_string):
-    """Testing getting the reference spectrum.
-
-    Currently "Alpha=" in the id-string is not implemented.
-    Currently "smpl=" in the id-string is not implemented.
-    """
-    with pytest.raises(NotImplementedError):
-        snrnorm.get_reference_spectrum(bad_string)
-
-
-def test_valid_snr_get_reference_spectrum():
-    """Testing getting the reference spectrum."""
-    ref_band = "J"
-    wav_ref, flux_ref = snrnorm.get_reference_spectrum(
-        "M0-K-1.0-100k", ref_band=ref_band
-    )
-    band_min, band_max = utils.band_limits(ref_band)
-
-    # Test the wavelength is in the reference band wavelength range
-    assert np.all(wav_ref <= band_max)
-    assert np.all(wav_ref >= band_min)
-
-    # test properties of output
-    assert len(wav_ref) == len(flux_ref)
-    assert isinstance(wav_ref, np.ndarray)
-    assert isinstance(flux_ref, np.ndarray)
-
-
-def test_get_reference_spectrum_in_nonexistent_file():
-    """Testing getting the reference spectrum."""
-    with pytest.raises(FileNotFoundError):
-        snrnorm.get_reference_spectrum("M1-K-1.0-100k", ref_band="J")
-
-
-def test_normalize_flux_new_verse_old(resampled_data):
-    """Test only small differences due to new normalization."""
-    id_string, wav, flux = resampled_data
-
-    print("wav in max =", wav[0], wav[-1])
-    new_norm = snrnorm.normalize_flux(flux, id_string, new=True)
-    old_norm = snrnorm.normalize_flux(flux, id_string, new=False)
-
-    print("new norm", new_norm)
-    print("old_norm", old_norm)
-
-    rvprec_new = Q.RVprec_calc(wav, new_norm)
-    rvprec_old = Q.RVprec_calc(wav, old_norm)
-
-    print("new rv=", rvprec_new, "old rv=", rvprec_old)
-    assert np.abs(rvprec_new.value - rvprec_old.value) < 0.4
-
-
-def test_old_normalization_does_not_handle_changed_band(resampled_data):
-    id_string, wav, flux = resampled_data
-    with pytest.raises(ValueError):
-        snrnorm.normalize_flux(flux, id_string, new=False, ref_band="K")
-
-
-def test_old_normalization_does_not_handle_changed_snr(resampled_data):
-    id_string, wav, flux = resampled_data
-    with pytest.raises(ValueError):
-        snrnorm.normalize_flux(flux, id_string, new=False, snr=101)
-
-
-@pytest.mark.parametrize(
-    "id_string",
-    [
-        "M0-1.0",
-        "M3-1.0",
-        "M6-1.0",
-        "M9-1.0",
-        "M0-5.0",
-        "M3-5.0",
-        "M6-5.0",
-        "M9-5.0",
-        "M0-10.0",
-        "M3-10.0",
-        "M6-10.0",
-        "M9-10.0",
-    ],
-)
-def test_snr_old_norm_constant(id_string):
-    norm_const = snrnorm.old_norm_constant(id_string)
-    assert isinstance(norm_const, float)
-
-
-@pytest.mark.parametrize(
-    "bad_string", ["M0-1", "M0-2.5", "M8-1.0", "M6-5", "M9-10", "T0-3.0", "", "AB-CDE"]
-)
-def test_snr_old_norm_constant_with_bad_id_str(bad_string):
-    """Fixed to the set of values in first paper."""
-    with pytest.raises(ValueError):
-        snrnorm.old_norm_constant(bad_string)
-
-
-def test_get_ref_spectrum_with_ref_band_self(resampled_data):
-    """Checks for upper or lower "self"."""
-    id_string, wav, flux = resampled_data
-
-    wav_ref, flux_ref = snrnorm.get_reference_spectrum(id_string, ref_band="self")
-
-    # Reference is the same values
-    assert np.allclose(wav, wav_ref)
-    assert np.allclose(flux, flux_ref)
-
-
-@pytest.mark.parametrize("ref_band", ["self", "SELF", "self", "SeLF"])
-def test_get_self_band_can_be_any_case(resampled_data, ref_band):
-    """Checks for upper or lower "self"."""
-
-    id_string, wav, flux = resampled_data
-    wav_ref, flux_ref = snrnorm.get_reference_spectrum(id_string, ref_band=ref_band)
-
-    # Reference is the same values
-    assert np.allclose(wav, wav_ref)
-    assert np.allclose(flux, flux_ref)
 
 
 @pytest.mark.parametrize("band", ["VIS", "Z", "NIR", "J", "Y"])
@@ -300,43 +149,6 @@ def test_snr_constant_band_with_invalid_wavelength(wav, band):
         snrnorm.snr_constant_band(wav, np.ones(50), band=band)
 
 
-@pytest.mark.parametrize(
-    "id_string",
-    [
-        "M0-BAD-1.0-100k",
-        "M9-A-5.0-50k",
-        "MO-J-1.0-100k",
-        "N0-J-1.0-100k",
-        "M2--1.0-100k",
-        "M0-J-2-100k",
-        "M9-Z-5.0",
-        "M0-J-1.0-100",
-        "M0-J-1.0-1k",
-        "M2-100k",
-        "M0",
-    ],
-)
-def test_decompose_bad_id_strings_give_errors(id_string):
-    with pytest.raises(ValueError):
-        snrnorm.decompose_id_string(id_string)
-
-
-@pytest.mark.parametrize(
-    "id_string,expected",
-    [
-        ("M0-H-1.0-100k", ("M0", "H", "1.0", "100k")),
-        ("M9-K-5.0-50k", ("M9", "K", "5.0", "50k")),
-        ("M9-J-5.0-30k", ("M9", "J", "5.0", "30k")),
-        ("M6-TEST-10.0-80k", ("M6", "TEST", "10.0", "80k")),
-    ],
-)
-def test_decompose_id_string(id_string, expected):
-    decomposed = snrnorm.decompose_id_string(id_string)
-
-    assert decomposed == expected
-    assert len(decomposed) == 4
-
-
 @pytest.mark.parametrize("wav_ref", [0.5, 4])
 def test_snr_constant_wav_ref_outside_wav(wav_ref):
     """Wav-ref outside bounds of wav should raise ValueError"""
@@ -345,7 +157,3 @@ def test_snr_constant_wav_ref_outside_wav(wav_ref):
 
     with pytest.raises(ValueError):
         snrnorm.snr_constant_wav(wav, flux, wav_ref)
-
-
-# TODO:
-#  Test normalize_flux() with ref_band == SELF. to check the condition on line 56
